lvl increment

+++ Assets/_FreedomForZombie/ScriptableObjects/ZombieSO.cs
@@ -8,7 +8,7 @@ public class ZombieSO : ScriptableObject
     [SerializeField] private string _descripription;
     [SerializeField] private int _defaultManaCost;
     [SerializeField] private int _defaultManyCost;
-    [SerializeField] private int _defaultDamage;
+    [SerializeField] private float _defaultDamage;
     [SerializeField] private int _defaultHealth;
     [SerializeField] private int _defaultAtackSpeed;
     [SerializeField] private float _defaultSpeed;
@@ -19,7 +19,7 @@ public class ZombieSO : ScriptableObject
     public string Descripription => _descripription;
     public int DefaultManaCost => _defaultManaCost;
     public int DefaultManyCost => _defaultManyCost;
-    public int DefaultDamage => _defaultDamage;
+    public float DefaultDamage => _defaultDamage;
     public int DefaultHealth => _defaultHealth;
     public int DefaultAtackSpeed => _defaultAtackSpeed;
     public float DefaultSpeed => _defaultSpeed;
